feat: split files better and add more components
