+++ components/chips/chip.test.mjs
@@ -1,0 +1,4 @@
+import { Chip } from './chip.mjs';
+
+const chip = new Chip().setText('John Doe');
+console.log(chip.toHTML());
