+++ components/lists/list.mjs
@@ -1,0 +1,11 @@
+import { Component } from '../component.mjs';
+
+class ListItem extends Component {
+  constructor(options) {
+    super(options);
+    //this.setTagName('div');
+    this.addClassname('list-item');
+  }
+}
+
+export { ListItem };
